requirements.txt
django
python-dotenv
openai
PyPDF2
requests
djangorestframework

